--- a/src/alg_utils.jl
+++ b/src/alg_utils.jl
@@ -1,4 +1,5 @@
-for order in (3, 4, 5, 6), alg in ((Symbol("GeneralMIRK$(order)"), Symbol("MIRK$(order)")))
+for order in (3, 4, 5, 6)
+    alg = Symbol("MIRK$(order)")
     @eval alg_order(::$(alg)) = $order
     @eval alg_stage(::$(alg)) = $(order - 1)
 end
@@ -7,4 +8,4 @@ SciMLBase.isautodifferentiable(::BoundaryValueDiffEqAlgorithm) = true
 SciMLBase.allows_arbitrary_number_types(::BoundaryValueDiffEqAlgorithm) = true
 SciMLBase.allowscomplex(alg::BoundaryValueDiffEqAlgorithm) = true
 
-SciMLBase.isadaptive(alg::Union{GeneralMIRK, MIRK}) = true
+SciMLBase.isadaptive(alg::AbstractMIRK) = true