--- a/src/algorithms.jl
+++ b/src/algorithms.jl
@@ -23,8 +23,23 @@ Shooting(ode_alg; nlsolve = DEFAULT_NLSOLVE_SHOOTING) = Shooting(ode_alg, nlsolv
   pages={479-497}
 }
 """
-struct GeneralMIRK3{N} <: GeneralMIRK
-    nlsolve::N
+Base.@kwdef struct GeneralMIRK3{N} <: GeneralMIRK
+    nlsolve::N = DEFAULT_NLSOLVE_MIRK
+end
+
+
+"""
+@article{Enright1996RungeKuttaSW,
+  title={Runge-Kutta Software with Defect Control for Boundary Value ODEs},
+  author={Wayne H. Enright and Paul H. Muir},
+  journal={SIAM J. Sci. Comput.},
+  year={1996},
+  volume={17},
+  pages={479-497}
+}
+"""
+Base.@kwdef struct GeneralMIRK4{N} <: GeneralMIRK
+    nlsolve::N = DEFAULT_NLSOLVE_MIRK
 end
 
 """
@@ -37,8 +52,8 @@ end
   pages={479-497}
 }
 """
-struct GeneralMIRK4{N} <: GeneralMIRK
-    nlsolve::N
+Base.@kwdef struct GeneralMIRK5{N} <: GeneralMIRK
+    nlsolve::N = DEFAULT_NLSOLVE_MIRK
 end
 
 """
@@ -51,8 +66,8 @@ end
   pages={479-497}
 }
 """
-struct GeneralMIRK5{N} <: GeneralMIRK
-    nlsolve::N
+Base.@kwdef struct GeneralMIRK6{N} <: GeneralMIRK
+    nlsolve::N = DEFAULT_NLSOLVE_MIRK
 end
 
 """
@@ -65,8 +80,8 @@ end
   pages={479-497}
 }
 """
-struct GeneralMIRK6{N} <: GeneralMIRK
-    nlsolve::N
+Base.@kwdef struct MIRK3{N} <: MIRK
+    nlsolve::N = DEFAULT_NLSOLVE_MIRK
 end
 
 """
@@ -79,8 +94,8 @@ end
   pages={479-497}
 }
 """
-struct MIRK3{N} <: MIRK
-    nlsolve::N
+Base.@kwdef struct MIRK4{N} <: MIRK
+    nlsolve::N = DEFAULT_NLSOLVE_MIRK
 end
 
 """
@@ -93,8 +108,8 @@ end
   pages={479-497}
 }
 """
-struct MIRK4{N} <: MIRK
-    nlsolve::N
+Base.@kwdef struct MIRK5{N} <: MIRK
+    nlsolve::N = DEFAULT_NLSOLVE_MIRK
 end
 
 """
@@ -107,29 +122,6 @@ end
   pages={479-497}
 }
 """
-struct MIRK5{N} <: MIRK
-    nlsolve::N
+Base.@kwdef struct MIRK6{N} <: MIRK
+    nlsolve::N = DEFAULT_NLSOLVE_MIRK
 end
-
-"""
-@article{Enright1996RungeKuttaSW,
-  title={Runge-Kutta Software with Defect Control for Boundary Value ODEs},
-  author={Wayne H. Enright and Paul H. Muir},
-  journal={SIAM J. Sci. Comput.},
-  year={1996},
-  volume={17},
-  pages={479-497}
-}
-"""
-struct MIRK6{N} <: MIRK
-    nlsolve::N
-end
-
-GeneralMIRK3(; nlsolve = DEFAULT_NLSOLVE_MIRK) = GeneralMIRK3(nlsolve)
-GeneralMIRK4(; nlsolve = DEFAULT_NLSOLVE_MIRK) = GeneralMIRK4(nlsolve)
-GeneralMIRK5(; nlsolve = DEFAULT_NLSOLVE_MIRK) = GeneralMIRK5(nlsolve)
-GeneralMIRK6(; nlsolve = DEFAULT_NLSOLVE_MIRK) = GeneralMIRK6(nlsolve)
-MIRK3(; nlsolve = DEFAULT_NLSOLVE_MIRK) = MIRK3(nlsolve)
-MIRK4(; nlsolve = DEFAULT_NLSOLVE_MIRK) = MIRK4(nlsolve)
-MIRK5(; nlsolve = DEFAULT_NLSOLVE_MIRK) = MIRK5(nlsolve)
-MIRK6(; nlsolve = DEFAULT_NLSOLVE_MIRK) = MIRK6(nlsolve)