--- a/src/cache.jl
+++ b/src/cache.jl
@@ -1,22 +1,20 @@
 abstract type AbstractMIRKCache end
 abstract type MIRKCache <: AbstractMIRKCache end
-abstract type GeneralMIRKCache <: AbstractMIRKCache end
 
 const AA3 = AbstractArray{T, 3} where {T}
 
 for order in (3, 4, 5, 6)
-    cache = Symbol("MIRK$(order)GeneralCache")
+    cache = Symbol("MIRK$(order)Cache")
     # `k_discrete` stores discrete stages for each subinterval,
     # hence the size of k_discrete is M × stage × (N - 1)
-    @eval struct $(cache){kType <: AA3} <: GeneralMIRKCache
+    @eval struct $(cache){kType <: AA3} <: MIRKCache
         k_discrete::kType
     end
 
     @eval @truncate_stacktrace $cache
 
-    for algType in (Symbol("GeneralMIRK$order"), Symbol("MIRK$order"))
-        @eval function alg_cache(::$algType, S::BVPSystem)
-            return $(cache)(similar(S.tmp, S.M, S.stage, S.N - 1))
-        end
+    algType = Symbol("MIRK$(order)")
+    @eval function alg_cache(::$algType, S::BVPSystem)
+        return $(cache)(similar(S.tmp, S.M, S.stage, S.N - 1))
     end
 end