--- a/src/jacobian.jl
+++ b/src/jacobian.jl
@@ -8,7 +8,7 @@ end
 #        this horror.
 function _construct_nonlinear_problem_with_jacobian(f!::BVPJacobianWrapper, S::BVPSystem,
     y, p)
-    J0 = BandedMatrix(similar(first(S.y), (S.M * S.N, S.M * S.N)), (S.M, S.M))[(S.M + 1):end,
+    J0 = BandedMatrix(similar(y, (S.M * S.N, S.M * S.N)), (S.M, S.M))[(S.M + 1):end,
         :]
     jac_cache = JacobianCache(similar(y),
         similar(y)[(S.M + 1):end];
@@ -30,6 +30,7 @@ function _construct_nonlinear_problem_with_jacobian(f!::BVPJacobianWrapper, S::B
             return
         end
         finite_difference_jacobian!(@view(J[begin:(S.M), :]), _f1!, x)
+        # display(sparse(J))
         return J
     end
     return NonlinearProblem(NonlinearFunction{true}(f!; jac = jac!), y, p)