--- a/src/solve.jl
+++ b/src/solve.jl
@@ -20,23 +20,16 @@ end
 function construct_MIRK_loss_function(S::BVPSystem, prob::BVProblem, TU, cache, mesh)
     function loss!(resid, u, p)
         u_ = reshape(u, S.M, S.N)
-        # @show 0, sum(u)
         resid_ = reshape(resid, S.M, S.N)
-        # @show 1, resid
         Φ!(resid_, S, TU, cache, u_, p, mesh)
-        # @show 1, resid, sum(u)
-        # @show 2, resid, u
         eval_bc_residual!(resid_, prob.problem_type, S, u_, p, mesh)
-        # @show 3, resid
-        # @show 2, resid, sum(resid) / length(resid), sum(u)
-        # error(1)
         return resid
     end
     return loss!
 end
 
-function DiffEqBase.__solve(prob::BVProblem, alg::Union{GeneralMIRK, MIRK}; dt = 0.0,
-    abstol = 1e-3, adaptive::Bool = true, kwargs...)
+function DiffEqBase.__solve(prob::BVProblem, alg::AbstractMIRK; dt = 0.0, abstol = 1e-3,
+    adaptive::Bool = true, kwargs...)
     dt ≤ 0 && throw(ArgumentError("dt must be positive"))
     T = eltype(prob.u0)
     n = Int(cld((prob.tspan[2] - prob.tspan[1]), dt))