# Benchmarks

## Old

```julia
julia> @benchmark sol = solve($prob, $GeneralMIRK4(), dt = 0.2)
BenchmarkTools.Trial: 1650 samples with 1 evaluation.
 Range (min … max):  1.771 ms … 56.913 ms  ┊ GC (min … max):  0.00% … 96.10%
 Time  (median):     2.902 ms              ┊ GC (median):     0.00%
 Time  (mean ± σ):   3.002 ms ±  3.446 ms  ┊ GC (mean ± σ):  11.86% ±  9.66%

    █                                                         
  █▃█▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂ ▂
  1.77 ms        Histogram: frequency by time        33.2 ms <

 Memory estimate: 3.58 MiB, allocs estimate: 45216.

julia> @benchmark sol = solve($prob, $GeneralMIRK5(), dt = 0.2)
BenchmarkTools.Trial: 1100 samples with 1 evaluation.
 Range (min … max):  2.537 ms … 62.432 ms  ┊ GC (min … max):  0.00% … 93.57%
 Time  (median):     4.141 ms              ┊ GC (median):     0.00%
 Time  (mean ± σ):   4.517 ms ±  4.885 ms  ┊ GC (mean ± σ):  12.66% ± 10.91%

  ▂ █                                                         
  █▃██▃▃▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▁▁▁▁▁▂▁▂ ▂
  2.54 ms        Histogram: frequency by time        37.2 ms <

 Memory estimate: 5.10 MiB, allocs estimate: 65094.

julia> @benchmark sol = solve($prob, $GeneralMIRK6(), dt = 0.2)
BenchmarkTools.Trial: 885 samples with 1 evaluation.
 Range (min … max):  3.326 ms … 60.302 ms  ┊ GC (min … max):  0.00% … 91.41%
 Time  (median):     5.485 ms              ┊ GC (median):     0.00%
 Time  (mean ± σ):   5.622 ms ±  4.882 ms  ┊ GC (mean ± σ):  11.86% ± 12.07%

  ▆▂ █▁                                                       
  █████▅▆▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▆ ▇
  3.33 ms      Histogram: log(frequency) by time       38 ms <

 Memory estimate: 6.87 MiB, allocs estimate: 88089.
```

## New

```julia
julia> @benchmark sol = solve($prob, $GeneralMIRK4(), dt = 0.2)
BenchmarkTools.Trial: 9662 samples with 1 evaluation.
 Range (min … max):  474.793 μs …  30.613 ms  ┊ GC (min … max): 0.00% … 97.84%
 Time  (median):     493.651 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   511.745 μs ± 509.445 μs  ┊ GC (mean ± σ):  2.46% ±  2.70%

   █▃    ▁                                                       
  ▇██▅▄▅██▇▆▄▄▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▁▂▂▂▂▁▁▁▂▂▁▁▂▂▁▂▂▂▁▂▂▂▂▂▂▂▂▂▂▂▂ ▃
  475 μs           Histogram: frequency by time          707 μs <

 Memory estimate: 172.95 KiB, allocs estimate: 1197.

julia> @benchmark sol = solve($prob, $GeneralMIRK5(), dt = 0.2)
BenchmarkTools.Trial: 7891 samples with 1 evaluation.
 Range (min … max):  561.399 μs …  33.515 ms  ┊ GC (min … max): 0.00% … 97.63%
 Time  (median):     590.663 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   626.672 μs ± 560.889 μs  ┊ GC (mean ± σ):  2.08% ±  2.61%

  ▇█▆▆▇▆▅▅▄▄▃▃▂▂▂▁▂▁ ▁                                          ▂
  ██████████████████████▇█▇▇▇▆▇▆▆▆▆▅▆▄▆▄▅▅▆▅▆▆▆▅▆▆▅▆▅▆▆▆▂▅▅▆▆▅▅ █
  561 μs        Histogram: log(frequency) by time        992 μs <

 Memory estimate: 181.02 KiB, allocs estimate: 1235.

julia> @benchmark sol = solve($prob, $GeneralMIRK6(), dt = 0.2)
BenchmarkTools.Trial: 7044 samples with 1 evaluation.
 Range (min … max):  591.711 μs …  31.098 ms  ┊ GC (min … max): 0.00% … 93.10%
 Time  (median):     633.686 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   695.969 μs ± 523.927 μs  ┊ GC (mean ± σ):  1.66% ±  2.40%

  █▇▇█▇▆▅▄▄▃▃▃▃▂▂▁▁▁▁                                           ▂
  ███████████████████▇█▆▇▆▇▆▆▇▆▅▅▅▅▅▆▆▆▆▆▅▆▆▇▆▄▇▇▇▇▇▇█▇█▇█▇▆▇█▆ █
  592 μs        Histogram: log(frequency) by time        1.3 ms <

 Memory estimate: 212.25 KiB, allocs estimate: 1321.
```