@avik-pal 2023-08-07
@ChrisRackauckas this is a go from my end. Main changes:
1. Drops General* methods since they don't do anything different.
2. Array of Arrays are restructured as Higher Dimensional Arrays to tap into BLAS calls directly.
3. The previous BVPSystem required unnecessary `copyto!`, I dropped most of the fields in that.